Extract user sync job from __init__
